+++ core/src/main/java/io/camunda/migrator/mapper/IdKeyMapper.java
@@ -5,8 +5,9 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.migrator.history;
+package io.camunda.migrator.mapper;
 
+import io.camunda.migrator.history.IdKeyDbModel;
 import java.util.Set;
 import org.apache.ibatis.annotations.Param;
